[authentication] named scanner broke auth factory
The authentication factory expected a fixed string for the scanners, to
get to the authentication type.

This was broken by the named scanner (e.g.: `zap_anonymous`).

This commit fixes that.

Also, while at it, I tidied a little:
- moved the factory into its own file
- improved the dev documentation for authentication
